char is 16 bits

+++ hash/index.js
@@ -18,11 +18,11 @@ const hash = 'WebAssembly' in global ?
             .then(module => { return new WebAssembly.Instance(module, {}); })
             .then(instance => {
                  return (seed = 0) => {
-                     return (str) => {                    
-                         const buf = new Uint8Array(instance.exports.memory.buffer);
+                     return (str) => {
+                         const buf = new Uint16Array(instance.exports.memory.buffer);
                          for(let i=0; i<str.length ;i++) {
                              buf[i] = str.charCodeAt(i);
-                         }                    
+                         }                         
                          return instance.exports.hash(seed) >>> 0;
                      };
                  };
